seed data is little weird
its not really matching up
but i made it work
